Renamed the ExecutorPluginConfig
