+++ client/driver/executor_plugin.go
@@ -1,51 +1,13 @@
 package driver
 
 import (
-	"io"
 	"log"
-	"net"
 	"net/rpc"
 
 	"github.com/hashicorp/go-plugin"
 	"github.com/hashicorp/nomad/client/driver/executor"
 	"github.com/hashicorp/nomad/nomad/structs"
 )
-
-var HandshakeConfig = plugin.HandshakeConfig{
-	ProtocolVersion:  1,
-	MagicCookieKey:   "NOMAD_PLUGIN_MAGIC_COOKIE",
-	MagicCookieValue: "e4327c2e01eabfd75a8a67adb114fb34a757d57eee7728d857a8cec6e91a7255",
-}
-
-func GetPluginMap(w io.Writer) map[string]plugin.Plugin {
-	p := new(ExecutorPlugin)
-	p.logger = log.New(w, "", log.LstdFlags)
-	return map[string]plugin.Plugin{"executor": p}
-}
-
-// ExecutorReattachConfig is the config that we seralize and de-serialize and
-// store in disk
-type ExecutorReattachConfig struct {
-	Pid      int
-	AddrNet  string
-	AddrName string
-}
-
-// PluginConfig returns a config from an ExecutorReattachConfig
-func (c *ExecutorReattachConfig) PluginConfig() *plugin.ReattachConfig {
-	var addr net.Addr
-	switch c.AddrNet {
-	case "unix", "unixgram", "unixpacket":
-		addr, _ = net.ResolveUnixAddr(c.AddrNet, c.AddrName)
-	case "tcp", "tcp4", "tcp6":
-		addr, _ = net.ResolveTCPAddr(c.AddrNet, c.AddrName)
-	}
-	return &plugin.ReattachConfig{Pid: c.Pid, Addr: addr}
-}
-
-func NewExecutorReattachConfig(c *plugin.ReattachConfig) *ExecutorReattachConfig {
-	return &ExecutorReattachConfig{Pid: c.Pid, AddrNet: c.Addr.Network(), AddrName: c.Addr.String()}
-}
 
 type ExecutorRPC struct {
 	client *rpc.Client
